+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # ChangeLog
 
+#### 4.6.7
+
+- fix: spelling error for 'nodeselectChange', closes: #3736;
+- fix: update node icon from show false to show true;
+- perf: update related edges while drag-combo;
+- feat: combo supports collapsedSubstituteIcon showing after collapsed;
+- feat: remove animations while first rendering with (collapsed)combos;
+
 #### 4.6.6
 
 - fix: destroyLayout error, closes: #3727;